requirements.txt
python-dotenv==1.0.0
openai>=1.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
pytz>=2023.3
python-dateutil>=2.8.0